fix(middleware): consolidate rate limiting, body size check, and simulator guard into src/middleware.ts
